feat: 프로필 화면의 사용자 로그아웃 및 탈퇴 기능 개선, UI 스타일 수정
