+++ src/app/(tabs)/(profile)/profile.tsx
@@ -1,7 +1,7 @@
 import { AlertIcon } from "@/assets/svgs/svgs";
-import { deleteUser, getUserCourses, invalidateToken } from "@/src/apis";
-import { GhostSortOption, UserCourseInfo } from "@/src/apis/types/course";
-import { CoursesWithFilter } from "@/src/components/course/CoursesWithFilter";
+import { deleteUser, invalidateToken } from "@/src/apis";
+import { UserCourseInfo } from "@/src/apis/types/course";
+import { CourseSection } from "@/src/components/profile/CourseSection";
 import { Info } from "@/src/components/profile/Info";
 import { ActionButton } from "@/src/components/ui/ActionButton";
 import BottomModal from "@/src/components/ui/BottomModal";
@@ -15,10 +15,9 @@ import { Typography } from "@/src/components/ui/Typography";
 import { useAuthStore } from "@/src/store/authState";
 import colors from "@/src/theme/colors";
 import { BottomSheetModal } from "@gorhom/bottom-sheet";
-import { useInfiniteQuery } from "@tanstack/react-query";
 import { useLocalSearchParams, useRouter } from "expo-router";
-import { useRef, useState } from "react";
-import { Alert, SafeAreaView, View } from "react-native";
+import { useCallback, useRef, useState } from "react";
+import { Alert, SafeAreaView, StyleSheet, View } from "react-native";
 import { ScrollView } from "react-native-gesture-handler";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 import Toast from "react-native-toast-message";
@@ -37,19 +36,43 @@ export default function ProfileScreen() {
     );
     const { bottom } = useSafeAreaInsets();
     const scrollViewRef = useRef<ScrollView>(null);
+
+    const onPressSignButton = useCallback(
+        async (modalType: "logout" | "withdraw") => {
+            bottomSheetRef.current?.close();
+            if (modalType === "logout") {
+                await invalidateToken();
+                Toast.show({
+                    type: "success",
+                    text1: "로그아웃 되었습니다.",
+                    position: "bottom",
+                });
+                logout();
+            } else {
+                Alert.alert("회원 탈퇴", "정말로 탈퇴하시겠습니까?", [
+                    {
+                        text: "취소",
+                        style: "cancel",
+                    },
+                    {
+                        text: "탈퇴",
+                        onPress: async () => {
+                            await deleteUser();
+                            logout();
+                        },
+                    },
+                ]);
+            }
+        },
+        [bottomSheetRef, logout]
+    );
     return (
-        <View style={{ flex: 1, backgroundColor: "#111111" }}>
-            <SafeAreaView
-                style={{
-                    flex: 1,
-                    backgroundColor: "#111111",
-                    marginBottom: bottom + 70,
-                }}
-            >
+        <View style={styles.container}>
+            <SafeAreaView style={styles.safeAreaView}>
                 {/* Header */}
                 <View>
                     <Header titleText="마이페이지" hasBackButton={false} />
-                    <View style={{ flexDirection: "row", marginTop: 10 }}>
+                    <View style={styles.header}>
                         <TabItem
                             title="내 정보"
                             onPress={() => setSelectedTab("info")}
@@ -71,7 +94,7 @@ export default function ProfileScreen() {
                     />
                 )}
                 {selectedTab === "course" && (
-                    <Course
+                    <CourseSection
                         selectedCourse={selectedCourse}
                         setSelectedCourse={setSelectedCourse}
                     />
@@ -81,21 +104,12 @@ export default function ProfileScreen() {
             <BottomModal
                 bottomSheetRef={bottomSheetRef}
                 canClose={true}
-                handleStyle={{
-                    paddingTop: 10,
-                    paddingBottom: 30,
-                }}
+                handleStyle={styles.handle}
             >
-                <View style={{ gap: 30 }}>
-                    <View
-                        style={{
-                            gap: 15,
-                            alignItems: "center",
-                            marginBottom: 30,
-                        }}
-                    >
+                <View style={styles.modalContainer}>
+                    <View style={styles.modalContent}>
                         <AlertIcon color={colors.red} />
-                        <View style={{ gap: 4, alignItems: "center" }}>
+                        <View style={styles.modalText}>
                             <Typography variant="headline" color="white">
                                 {modalType === "logout"
                                     ? "로그아웃 하시겠습니까?"
@@ -110,36 +124,7 @@ export default function ProfileScreen() {
                     </View>
                     <ButtonWithIcon
                         iconType="home"
-                        onPress={async () => {
-                            bottomSheetRef.current?.close();
-                            if (modalType === "logout") {
-                                await invalidateToken();
-                                Toast.show({
-                                    type: "success",
-                                    text1: "로그아웃 되었습니다.",
-                                    position: "bottom",
-                                });
-                                logout();
-                            } else {
-                                Alert.alert(
-                                    "회원 탈퇴",
-                                    "정말로 탈퇴하시겠습니까?",
-                                    [
-                                        {
-                                            text: "취소",
-                                            style: "cancel",
-                                        },
-                                        {
-                                            text: "탈퇴",
-                                            onPress: async () => {
-                                                await deleteUser();
-                                                logout();
-                                            },
-                                        },
-                                    ]
-                                );
-                            }
-                        }}
+                        onPress={() => onPressSignButton(modalType)}
                         onPressIcon={() => {
                             router.replace("/");
                         }}
@@ -178,47 +163,34 @@ export default function ProfileScreen() {
     );
 }
 
-const Course = ({
-    selectedCourse,
-    setSelectedCourse,
-}: {
-    selectedCourse: UserCourseInfo | null;
-    setSelectedCourse: (course: UserCourseInfo | null) => void;
-}) => {
-    const { data, isLoading, isError, fetchNextPage, hasNextPage } =
-        useUserCourses();
-
-    if (isLoading) {
-        return <></>;
-    }
-    if (isError) {
-        return <Typography>코스 정보를 불러오는데 실패했습니다.</Typography>;
-    }
-    return (
-        <View style={{ marginTop: 20, flex: 1 }}>
-            <CoursesWithFilter
-                data={data?.pages.flatMap((page) => page.content) ?? []}
-                selectedCourse={selectedCourse}
-                setSelectedCourse={setSelectedCourse}
-                hasNextPage={hasNextPage}
-                fetchNextPage={fetchNextPage}
-            />
-        </View>
-    );
-};
-
-export function useUserCourses(
-    size: number = 10,
-    sort: GhostSortOption = "id,asc"
-) {
-    return useInfiniteQuery({
-        queryKey: ["user-courses", size, sort],
-        queryFn: async ({ pageParam = 0 }) =>
-            getUserCourses({ page: pageParam, size, sort }),
-        getNextPageParam: (lastPage) => {
-            const { number, totalPages } = lastPage.page;
-            return number + 1 < totalPages ? number + 1 : undefined;
-        },
-        initialPageParam: 0,
-    });
-}
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: "#111111",
+    },
+    safeAreaView: {
+        flex: 1,
+        backgroundColor: "#111111",
+        marginBottom: 70,
+    },
+    header: {
+        marginTop: 10,
+        flexDirection: "row",
+    },
+    handle: {
+        paddingTop: 10,
+        paddingBottom: 30,
+    },
+    modalContainer: {
+        gap: 30,
+    },
+    modalContent: {
+        gap: 15,
+        alignItems: "center",
+        marginBottom: 30,
+    },
+    modalText: {
+        gap: 4,
+        alignItems: "center",
+    },
+});
